refactor(useMemoryStore): extract display types and helper functions (#147)
